Move Collector for assets to asset Namespace
